Early support for "SpriteTechnique" style shaders in the utility technique delegate - no technique delegate specific patches yet

+++ RenderCore/LightingEngine/DeferredLightingResolve.cpp
@@ -387,7 +387,7 @@ namespace RenderCore { namespace LightingEngine
 		geoInitBuffer.resize((sphereGeo.size() + cubeGeo.size()) * sizeof(Float3));
 		std::memcpy(geoInitBuffer.data(), cubeGeo.data(), cubeGeo.size() * sizeof(Float3));
 		std::memcpy(PtrAdd(geoInitBuffer.data(), cubeGeo.size() * sizeof(Float3)), sphereGeo.data(), sphereGeo.size() * sizeof(Float3));
-		_geo = device.CreateResource(CreateDesc(BindFlag::VertexBuffer|BindFlag::TransferDst, LinearBufferDesc::Create(geoInitBuffer.size())), "light-stenciling-geometry");
+		_geo = device.CreateResource(CreateDesc(BindFlag::VertexBuffer|BindFlag::TransferDst, LinearBufferDesc::Create((unsigned)geoInitBuffer.size())), "light-stenciling-geometry");
 		_pendingGeoInitBuffer = std::move(geoInitBuffer);
 		_cubeOffsetAndCount = {0u, (unsigned)cubeGeo.size()};
 		_sphereOffsetAndCount = {(unsigned)cubeGeo.size(), (unsigned)sphereGeo.size()};
@@ -399,11 +399,11 @@ namespace RenderCore { namespace LightingEngine
 		//								= sin(pi/3)*cos(pi/6);
 		float underestimationFactor = std::sin(gPI/3.0f) * std::cos(gPI/6.0f);
 		for (auto& pt:lowDetailHemi.second) pt /= underestimationFactor;
-		_lowDetailHemiSphereVB = device.CreateResource(CreateDesc(BindFlag::VertexBuffer|BindFlag::TransferDst, LinearBufferDesc::Create(sizeof(decltype(lowDetailHemi.second)::value_type)*lowDetailHemi.second.size())), "light-stenciling-geometry");
+		_lowDetailHemiSphereVB = device.CreateResource(CreateDesc(BindFlag::VertexBuffer|BindFlag::TransferDst, LinearBufferDesc::Create(unsigned(sizeof(decltype(lowDetailHemi.second)::value_type)*lowDetailHemi.second.size()))), "light-stenciling-geometry");
 		_pendingLowDetailHemisphereVB = std::move(lowDetailHemi.second);
 		std::vector<uint16_t> ib; ib.reserve(lowDetailHemi.first.size());
 		std::copy(lowDetailHemi.first.begin(), lowDetailHemi.first.end(), std::back_inserter(ib));
-		_lowDetailHemiSphereIB = device.CreateResource(CreateDesc(BindFlag::IndexBuffer|BindFlag::TransferDst, LinearBufferDesc::Create(sizeof(uint16_t)*ib.size())), "light-stenciling-geometry");
+		_lowDetailHemiSphereIB = device.CreateResource(CreateDesc(BindFlag::IndexBuffer|BindFlag::TransferDst, LinearBufferDesc::Create(unsigned(sizeof(uint16_t)*ib.size()))), "light-stenciling-geometry");
 		_lowDetailHemiSphereIndexCount = (unsigned)ib.size();
 		_pendingLowDetailHemisphereIB = std::move(ib);
 	}
